+++ src/main/java/com/inditex/challenge/controller/PriceController.java
@@ -2,6 +2,7 @@ package com.inditex.challenge.controller;
 
 import com.inditex.challenge.dto.PriceResultDTO;
 import com.inditex.challenge.service.impl.PriceServiceImpl;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,11 +19,11 @@ public class PriceController {
     }
 
     @GetMapping("/price")
-    List<PriceResultDTO> getProducts(
-            @RequestParam String applyDate,
-            @RequestParam Integer productId,
-            @RequestParam Integer chainId
-    ) {
+    PriceResultDTO getProducts(
+            @RequestParam(name = "apply_date") String applyDate,
+            @RequestParam(name = "product_id") Integer productId,
+            @RequestParam(name = "chain_id") Integer chainId
+    ) throws Exception {
         // In case the number of parameters is higher, then we need to use a Wrapper. Because passing too many arguments to a method it will be end as a Code Smell
         priceService.validate(applyDate, productId, chainId);
         return priceService.process(applyDate, productId, chainId);