+++ src/main/java/com/inditex/challenge/exception/ErrorCode.java
@@ -1,0 +1,7 @@
+package com.inditex.challenge.exception;
+
+public class ErrorCode {
+
+    final public static String NEGATIVE_ID = "Id should be positive.";
+    final public static String BAD_DATE_FORMAT = "Date format should be YYYY-MM-DD";
+}