+++ CHANGELOG.md
@@ -31,7 +31,7 @@ Dates in this file will have the following format: MM/DD/YYYY
 - More logic in PriceConverter to convert records to response
 - Exception package created to handle exceptions
 - Logic with records in service to get final response. Search max priority record, find records in repository, convert records in response
-- 
+- Creation of query in repository to get the records that we are looking for
 
 ## [0.3.0] - 03/09/2022 - Branch: feature/price-endpoint